replace int in range proof parameters with uint

+++ token/core/zkatdlog/crypto/rp/ipa_test.go
@@ -21,7 +21,7 @@ var _ = Describe("Inner Product Argument", func() {
 		Context("If IPA is generated correctly", func() {
 			It("Succeeds", func() {
 				curve := math.Curves[0]
-				nr := 6
+				nr := uint(6)
 				l := int(math2.Pow(2, float64(nr)))
 				leftGens := make([]*math.G1, l)
 				rightGens := make([]*math.G1, l)
